+++ src/routes/index.js
@@ -1,9 +1,11 @@
+// Core import
 const express = require('express');
 const router = express.Router();
 
-router.get('/nextMatch', (req, res, next) => {
-  res.send('this is the next match');
-});
+// Route import
+const nextMatchRouter = require('./nextMatch.routes'); 
+
+router.use('/nextMatch', nextMatchRouter);
 
 router.get('/team', (req, res, next) => {
   res.send('this is the team');