+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+# 1.1.0
+
+# Add
+- Add controllers 'nextMatch'
+- Change router logic
+
 # 1.0.0
 
 # Add
 - Add index ms
 - Install express, lodash, mysql2
-- Add routes
+- Add routes 'nextMatch' and 'team'